+++ include/internal/iutest_charcode.hpp
@@ -18,10 +18,10 @@
 //======================================================================
 // include
 #include "iutest_port.hpp"
+#include "iutest_locale.hpp"
 #include "iutest_constant.hpp"
 
 #if IUTEST_HAS_CXX_HDR_CODECVT
-#  include <locale>
 #  include <codecvt>
 #endif
 
@@ -60,6 +60,26 @@ inline ::std::string AnyStringToMultiByteString(const char* str, int num = -1)
  * @return  マルチバイト文字列
 */
 ::std::string AnyStringToMultiByteString(const wchar_t* str, int num=-1);
+
+#if IUTEST_HAS_CHAR8_T
+
+/**
+ * @brief   ワイド文字列からマルチバイトへ変換
+ * @param [in]  str = 入力
+ * @param [in]  num = 入力バッファサイズ
+ * @return  マルチバイト文字列
+*/
+::std::string AnyStringToUTF8(const char8_t* str, int num = -1);
+
+/**
+ * @brief   ワイド文字列からマルチバイトへ変換
+ * @param [in]  str = 入力
+ * @param [in]  num = 入力バッファサイズ
+ * @return  マルチバイト文字列
+*/
+::std::string AnyStringToMultiByteString(const char8_t* str, int num = -1);
+
+#endif
 
 #if IUTEST_HAS_CHAR16_T
 
@@ -128,6 +148,7 @@ template<typename CharType>
     {
         return kStrings::Null;
     }
+    ScopedEncoding(LC_CTYPE, "UTF-8");
     return AnyStringToMultiByteString(any_c_str);
 }